feat: add new module to group datasource things
