--- app/src/main/java/wottrich/github/io/githubprofile/data/resource/ApiExceptions.kt
@@ -1,12 +1,0 @@
-package wottrich.github.io.githubprofile.data.resource
-
-/**
- * @author Wottrich
- * @author wottrich78@gmail.com
- * @since 22/09/2021
- *
- * Copyright © 2021 GithubProfile. All rights reserved.
- *
- */
-
-class EmptyDataException : Exception("We don't support null data on success state")
