+++ app/src/main/java/wottrich/github/io/githubprofile/view/widgets/HeaderProfile.kt
@@ -11,12 +11,12 @@ import androidx.compose.ui.unit.dp
 import coil.compose.LocalImageLoader
 import coil.compose.rememberImagePainter
 import coil.transform.CircleCropTransformation
+import github.io.wottrich.datasource.models.Profile
 import github.io.wottrich.ui.values.Description
 import github.io.wottrich.ui.values.Subtitle
 import github.io.wottrich.ui.values.Title
 import github.io.wottrich.ui.widgets.TextView
 import wottrich.github.io.githubprofile.R
-import wottrich.github.io.githubprofile.model.Profile
 
 /**
  * @author Wottrich
