+++ app/src/main/java/wottrich/github/io/githubprofile/viewModel/ProfileViewModel.kt
@@ -1,5 +1,9 @@
 package wottrich.github.io.githubprofile.viewModel
 
+import github.io.wottrich.datasource.GithubDataSourceInterface
+import github.io.wottrich.datasource.dispatchers.AppDispatchers
+import github.io.wottrich.datasource.models.Profile
+import github.io.wottrich.datasource.models.Repository
 import github.io.wottrich.ui.state.State
 import github.io.wottrich.ui.state.StateInitial
 import kotlinx.coroutines.flow.MutableStateFlow
@@ -7,10 +11,6 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.collect
 import wottrich.github.io.githubprofile.R
 import wottrich.github.io.githubprofile.archive.toState
-import wottrich.github.io.githubprofile.data.datasource.GithubDataSourceInterface
-import wottrich.github.io.githubprofile.model.Profile
-import wottrich.github.io.githubprofile.model.Repository
-import wottrich.github.io.githubprofile.util.AppDispatchers
 
 /**
  * @author Wottrich
