+++ build.gradle.kts
@@ -15,8 +15,7 @@ buildscript {
         }
     }
     dependencies {
-        classpath("com.android.tools.build:gradle:${Versions.gradleVersion}")
-        classpath("org.jetbrains.kotlin:kotlin-gradle-plugin:${Versions.kotlinVersion}")
+        defaultClasspath()
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
