+++ buildSrc/src/main/kotlin/Plugins.kt
@@ -1,0 +1,6 @@
+object Plugins {
+
+    const val androidLibrary = "com.android.library"
+    const val kotlinAndroid = "kotlin-android"
+
+}
