+++ datasource/src/main/AndroidManifest.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8"?>
+<manifest package="github.io.wottrich.datasource">
+
+</manifest>
