+++ datasource/src/main/java/github/io/wottrich/datasource/resource/ApiExceptions.kt
@@ -1,0 +1,12 @@
+package github.io.wottrich.datasource.resource
+
+/**
+ * @author Wottrich
+ * @author wottrich78@gmail.com
+ * @since 22/09/2021
+ *
+ * Copyright © 2021 GithubProfile. All rights reserved.
+ *
+ */
+
+class EmptyDataException : Exception("We don't support null data on success state")
